Buy and sell market ledger balance update added.

+++ routes/orders.js
@@ -271,9 +271,16 @@ const marketOrders = async() => {
            
     }else if(direction === "sell"){
       const marketSell = await db.query('SELECT * FROM marketOrders WHERE direction = ? ORDER BY time ASC;',"sell")
-      const updateBids = await db.query('select size,price from bidLimitOrders ORDER BY price DESC,time ASC;')
+      const updateBids = await db.query('select size,price,userId from bidLimitOrders ORDER BY price DESC,time ASC;')
       const marketSellArr = marketSell[0];
       const updateBidsArr = updateBids[0];
+
+
+      //fetch wallet balance
+
+      const walletBalance = await db.query("select usd,bitcoin from ledger where userId = ?",[userId]);
+      const walletBalanceArr = walletBalance[0];
+
 
       let limitSize = updateBidsArr[0].size;
       let index = 1;
@@ -302,11 +309,26 @@ const marketOrders = async() => {
     if(sub !== 0){
         
         if(marketSellArr[0].size > limitSize){
-            const avgPrice = costBasis / limitSize;
-            const queryStringTrades = "insert into trades(symbol,direction,price,size,time,userId) values(?,?,?,?,?,?);"
+            
+            if(walletBalanceArr[0].bitcoin >= limitSize){
 
-            const deleteRows = await db.query('DELETE FROM bidLimitOrders ORDER BY price DESC,time ASC LIMIT ?;',index);
-            if(deleteRows[0].affectedRows !== 0){
+                for(let y = 0; y < index; y++){
+
+                    const fetchLimitWallet = await db.query("select bitcoin from ledger where userId = ?;",[updateBidsArr[y].userId]);
+                    const updateBtcVal = fetchLimitWallet[0][0].bitcoin + updateBidsArr[y].size;
+                    const updateWallet = await db.query("update ledger set bitcoin = ? where userId = ?;",[updateBtcVal,updateBidsArr[y].userId]);
+                }
+
+                const avgPrice = costBasis / limitSize;
+
+                const updateBtcVal = walletBalanceArr[0].bitcoin - limitSize - (limitSize/100 * feesTaker);
+                const updateUsdVal = walletBalanceArr[0].usd + avgPrice * limitSize;
+                const updateWallet = await db.query("update ledger set bitcoin = ?, usd = ? where userId = ?;",[updateBtcVal,updateUsdVal,userId]);
+
+                const queryStringTrades = "insert into trades(symbol,direction,price,size,time,userId) values(?,?,?,?,?,?);"
+
+                const deleteRows = await db.query('DELETE FROM bidLimitOrders ORDER BY price DESC,time ASC LIMIT ?;',index);
+                if(deleteRows[0].affectedRows !== 0){
                 
                 const tradesInsert = await db.query(queryStringTrades,[symbol,direction,avgPrice,limitSize,time,userId]);
                 
@@ -315,17 +337,36 @@ const marketOrders = async() => {
                 const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
                 res.status(200).json({message : "Market Order Partially Filled Successfully", code : 200})
 
+                }
+
+            }else{
+                const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
+                res.status(403).json({message : "Not enough balance to fulfil this order.",code : 403});
             }  
 
         }else if(limitSize > marketSellArr[0].size){
             if(index === 1){
-                const avgPrice = updateBidsArr[0].price;
-                const queryStringTrades = "insert into trades(symbol,direction,price,size,time,userId) values(?,?,?,?,?,?);"
-            
-                const updateRow = await db.query('UPDATE bidLimitOrders SET size = ? ORDER BY price DESC,time ASC LIMIT ?',[sub,1]);
                 
-                
-                if(updateRow[0].changedRows !== 0){
+                if(walletBalanceArr[0].bitcoin >= size){
+
+                    const fetchWalletLimit = await db.query("select bitcoin from ledger where userId = ?;",[updateBidsArr[0].userId]);
+                    const updateBtcVal = fetchWalletLimit[0][0].bitcoin + sub;
+                    const updateWalletLimit = await db.query("update ledger set bitcoin = ? where userId = ?;",[updateBtcVal,updateBidsArr[0].userId]);
+                    
+                    if(updateWalletLimit[0].changedRows !== 0){
+                        const avgPrice = updateBidsArr[0].price;
+
+                        const queryStringTrades = "insert into trades(symbol,direction,price,size,time,userId) values(?,?,?,?,?,?);"
+                    
+                        const updateRow = await db.query('UPDATE bidLimitOrders SET size = ? ORDER BY price DESC,time ASC LIMIT ?',[sub,1]);
+                    
+                    
+                        if(updateRow[0].changedRows !== 0){
+                        
+                        const updateValBtc = walletBalanceArr[0].bitcoin - size - (size/100 * feesTaker);
+                        const updateValUsd = walletBalanceArr[0].usd + avgPrice * size;
+                        const updateWallet = await db.query("update ledger set bitcoin = ?, usd = ? where userId = ?; ",[updateValBtc,updateValUsd,userId]);
+                        
                     const tradesInsert = await db.query(queryStringTrades,[symbol,direction,avgPrice,size,time,userId]);
                   
                     //Deleting market order from the database after the trade execution 
@@ -333,13 +374,40 @@ const marketOrders = async() => {
                     const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
                     res.status(200).json({message : "Market Order Filled Successfully", code : 200})
                     
+                        }
+                    }  
+
+                }else{
+                    const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
+                    res.status(403).json({message : "Not enough balance to fulfil this order.",code : 403});
                 }
+
             }else{
-                const rows = index - 1
+                const rows = index - 1;
+
+                const bidSub = costBasis - updateBidsArr[rows].price * sub;
+                const avgPrice = bidSub / marketSellArr[0].size;
+                console.log(avgPrice);
+
+                if(walletBalanceArr[0].bitcoin >= size){
+
+                    for(let a = 0; a < rows; a++){
+
+                        const fetchLimitWallet = await db.query("select bitcoin from ledger where userId = ?;",[updateBidsArr[a].userId]);
+                        const updateBtcVal = fetchLimitWallet[0][0].bitcoin + updateBidsArr[a].size;
+                        const updateWallet = await db.query("update ledger set bitcoin = ? where userId = ?;",[updateBtcVal,updateBidsArr[a].userId]);
+
+                    }
+                   
+                    const fetchLimitWallet = await db.query("select bitcoin from ledger where userId = ?;",[updateBidsArr[rows].userId]);
+                    const updateBtcVal = fetchLimitWallet[0][0].bitcoin + sub;
+                    const updateWallet = await db.query("update ledger set bitcoin = ? where userId = ?;",[updateBtcVal,updateBidsArr[rows].userId]);
+
+                   if(updateWallet[0].changedRows !== 0){
+
+                       const deleteRows = await db.query('DELETE FROM bidLimitOrders ORDER BY price DESC,time ASC LIMIT ?;',rows);
                 
-                const deleteRows = await db.query('DELETE FROM bidLimitOrders ORDER BY price DESC,time ASC LIMIT ?;',rows);
-                
-                if(deleteRows[0].affectedRows !== 0){
+                       if(deleteRows[0].affectedRows !== 0){
                     const queryStringTrades = "insert into trades(symbol,direction,price,size,time,userId) values(?,?,?,?,?,?);"
                     
                     const updateRow = await db.query('UPDATE bidLimitOrders SET size = ? ORDER BY price DESC,time ASC LIMIT ?',[sub,1]);
@@ -347,9 +415,13 @@ const marketOrders = async() => {
                     
                     
                     if(updateRow[0].changedRows !== 0){
-                        const fetchBids = await db.query('select size,price from bidLimitOrders ORDER BY price DESC,time ASC;');
-                        const bidSub = costBasis - fetchBids[0][0].price * fetchBids[0][0].size 
-                        const avgPrice =  bidSub / marketSellArr[0].size;
+                        // const fetchBids = await db.query('select size,price from bidLimitOrders ORDER BY price DESC,time ASC;');
+                        // const bidSub = costBasis - fetchBids[0][0].price * fetchBids[0][0].size 
+                        // const avgPrice =  bidSub / marketSellArr[0].size;
+                        const updateBtcValue = walletBalanceArr[0].bitcoin - size - (size/100 * feesTaker);
+                        const updateUsdValue = walletBalanceArr[0].usd + avgPrice * size;
+                        const updateWallet = await db.query("update ledger set bitcoin = ?, usd = ? where userId = ?;",[updateBtcValue,updateUsdValue,userId]);
+
                         const tradesInsert = await db.query(queryStringTrades,[symbol,direction,avgPrice,size,time,userId]);
                         
                        //Deleting market order from the database after the trade execution 
@@ -360,9 +432,15 @@ const marketOrders = async() => {
 
                     }
                     
+                       }
+
+                    }
+
+                }else{
+                    const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
+                    res.status(403).json({message : "Not enough balance to fulfil this order.",code : 403});
                 }
-                
-                
+
             }
             
         }
@@ -371,21 +449,44 @@ const marketOrders = async() => {
         
         
     }else if(sub === 0){
-        const avgPrice = costBasis / limitSize;
-        const deleteRows = await db.query('DELETE FROM bidLimitOrders ORDER BY price DESC,time ASC LIMIT ?;',index);
-        
-        const queryStringTrades = "insert into trades(symbol,direction,price,size,time,userId) values(?,?,?,?,?,?);"
-        
-        
-        if(deleteRows[0].affectedRows !== 0){
-            const tradesInsert = await db.query(queryStringTrades,[symbol,direction,avgPrice,size,time,userId])
-            
-            //Deleting market order from the database after the trade execution 
-            
-            const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
-            res.status(200).json({message : "Market Order Filled Successfully", code : 200})
 
+        if(walletBalanceArr[0].bitcoin >= size){
+
+            for(let y = 0; y < index; y++){
+
+                const fetchLimitWallet = await db.query("select bitcoin from ledger where userId = ?;",[updateBidsArr[y].userId]);
+                const updateBtcVal = fetchLimitWallet[0][0].bitcoin + updateBidsArr[y].size;
+                const updateWallet = await db.query("update ledger set bitcoin = ? where userId = ?;",[updateBtcVal,updateBidsArr[y].userId]);
+
+            };
+
+            const avgPrice = costBasis / limitSize;
+            
+            const deleteRows = await db.query('DELETE FROM bidLimitOrders ORDER BY price DESC,time ASC LIMIT ?;',index);
+            
+            const queryStringTrades = "insert into trades(symbol,direction,price,size,time,userId) values(?,?,?,?,?,?);"
+            
+            
+            if(deleteRows[0].affectedRows !== 0){
+                
+                const updateBtcValue = walletBalanceArr[0].bitcoin - size - (size/100 * feesTaker);
+                const updateUsdValue = walletBalanceArr[0].usd + avgPrice * size;
+                const updateWallet = await db.query("update ledger set bitcoin = ?, usd = ? where userId = ?;",[updateBtcValue,updateUsdValue,userId]);
+
+                const tradesInsert = await db.query(queryStringTrades,[symbol,direction,avgPrice,size,time,userId])
+            
+                //Deleting market order from the database after the trade execution 
+            
+                const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
+                res.status(200).json({message : "Market Order Filled Successfully", code : 200})
+
+            }
+
+        }else{
+            const deleteRows = await db.query('DELETE FROM marketOrders ORDER BY time ASC LIMIT ?;',1);
+            res.status(403).json({message : "Not enough balance to fulfil this order.",code : 403});
         }
+
     }
